requirements.txt
pybluez
openpyxl
pyserial
